Fixed tag bug
